Wms dev

+++ API_WMS/miebach/rest_api/views/inbound.py
@@ -1960,7 +1960,9 @@ def get_purchaseorder_locations(put_zone, temp_dict):
     seller_id = temp_dict.get('seller_id', '')
     location_masters = LocationMaster.objects.filter(zone__user=user).exclude(
         lock_status__in=['Inbound', 'Inbound and Outbound'])
-    exclude_zones_list = ['QC_ZONE', 'DAMAGED_ZONE', 'RTO_ZONE', 'Non Sellable Zone']
+    exclude_zones_list = get_exclude_zones(User.objects.get(id=user))
+    exclude_zones_list.append('RTO_ZONE')
+    # exclude_zones_list = ['QC_ZONE', 'DAMAGED_ZONE', 'RTO_ZONE', 'Non Sellable Zone']
     if put_zone in exclude_zones_list:
         location = location_masters.filter(zone__zone=put_zone, zone__user=user)
         if location:
@@ -2216,6 +2218,8 @@ def save_po_location(put_zone, temp_dict, seller_received_list=[], run_segregati
         seller_received_list = [{'seller_id': '', 'sku_id': (purchase_data['sku']).id,
                                 'quantity': received_quantity, 'id': ''}]
     for po_received in seller_received_list:
+        if po_received.get('put_zone', ''):
+            put_zone = po_received['put_zone']
         temp_dict['seller_id'] = po_received.get('seller_id', '')
         batch_dict['transact_type'] = 'po'
         batch_dict['transact_id'] = data.id
@@ -2223,7 +2227,8 @@ def save_po_location(put_zone, temp_dict, seller_received_list=[], run_segregati
         if sellable_segregation == 'true' and run_segregation:
             if batch_obj and po_received.get('id', ''):
                 SellerPOSummary.objects.filter(id=po_received['id']).update(batch_detail_id=batch_obj.id)
-            create_update_primary_segregation(data, po_received['quantity'], temp_dict, batch_obj=batch_obj)
+            create_update_primary_segregation(data, po_received['quantity'], temp_dict, batch_obj=batch_obj,
+                                              sps_id=po_received.get('id', ''))
             continue
         location = get_purchaseorder_locations(put_zone, temp_dict)
         received_quantity = po_received['quantity']
@@ -2406,9 +2411,12 @@ def get_seller_receipt_id(purchase_order):
     return receipt_number
 
 
-def create_update_primary_segregation(data, quantity, temp_dict, batch_obj=None):
+def create_update_primary_segregation(data, quantity, temp_dict, batch_obj=None, sps_id=''):
     if not batch_obj and not 'quality_check' in temp_dict.keys():
-        segregation_obj = PrimarySegregation.objects.filter(purchase_order_id=data.id)
+        primary_filt_dict = {'purchase_order_id': data.id}
+        if sps_id:
+            primary_filt_dict['seller_po_summary_id'] = sps_id
+        segregation_obj = PrimarySegregation.objects.filter(**primary_filt_dict)
         if segregation_obj:
             segregation_obj = segregation_obj[0]
             segregation_obj.quantity = float(segregation_obj.quantity) + quantity
@@ -2416,12 +2424,17 @@ def create_update_primary_segregation(data, quantity, temp_dict, batch_obj=None)
                 segregation_obj.status = 1
             segregation_obj.save()
         else:
-            segregation_obj = PrimarySegregation.objects.create(purchase_order_id=data.id, quantity=quantity,status=1,
-                                              creation_date=datetime.datetime.now())
+            primary_seg_dict = {'purchase_order_id': data.id, 'quantity': quantity, 'status': 1,
+                                'creation_date': datetime.datetime.now()}
+            if sps_id:
+                primary_seg_dict['seller_po_summary_id'] = sps_id
+            segregation_obj = PrimarySegregation.objects.create(**primary_seg_dict)
     else:
         if batch_obj:
-            segregation_obj = PrimarySegregation.objects.filter(purchase_order_id=data.id,
-                                                                batch_detail_id=batch_obj.id)
+            primary_filt_dict = {'purchase_order_id': data.id, 'batch_detail_id': batch_obj.id}
+            if sps_id:
+                primary_filt_dict['seller_po_summary_id'] = sps_id
+            segregation_obj = PrimarySegregation.objects.filter(**primary_filt_dict)
             if segregation_obj:
                 segregation_obj = segregation_obj[0]
                 segregation_obj.quantity = float(segregation_obj.quantity) + quantity
@@ -2429,9 +2442,12 @@ def create_update_primary_segregation(data, quantity, temp_dict, batch_obj=None)
                     segregation_obj.status = 1
                 segregation_obj.save()
             else:
-                segregation_obj = PrimarySegregation.objects.create(purchase_order_id=data.id, quantity=quantity, status=1,
-                                                                    creation_date=datetime.datetime.now(),
-                                                                    batch_detail_id=batch_obj.id)
+                primary_seg_dict = {'purchase_order_id': data.id, 'quantity': quantity, 'status': 1,
+                                    'creation_date': datetime.datetime.now(),
+                                    'batch_detail_id': batch_obj.id}
+                if sps_id:
+                    primary_seg_dict['seller_po_summary_id'] = sps_id
+                segregation_obj = PrimarySegregation.objects.create(**primary_seg_dict)
 
 def update_seller_po(data, value, user, myDict, i, receipt_id='', invoice_number='', invoice_date=None,
                      challan_number='', challan_date=None, dc_level_grn='', round_off_total=0):
@@ -2459,6 +2475,14 @@ def update_seller_po(data, value, user, myDict, i, receipt_id='', invoice_number
     overall_discount = 0
     if 'overall_discount' in myDict.keys() and myDict['overall_discount'][0]:
         overall_discount = myDict['overall_discount'][0]
+    remarks_list = []
+    if float(data.open_po.mrp) != float(myDict['mrp'][i]) :
+         remarks_list.append("mrp_change")
+    if 'offer_applicable' in myDict.keys() :
+        offer_applicable = myDict['offer_applicable'][i]
+        if offer_applicable == 'true':
+            remarks_list.append("offer_applied")
+    remarks = ','.join(remarks_list)
     if user.userprofile.user_type == 'warehouse_user':
         seller_po_summary, created = SellerPOSummary.objects.get_or_create(receipt_number=receipt_id,
                                                                            invoice_number=invoice_number,
@@ -2473,10 +2497,11 @@ def update_seller_po(data, value, user, myDict, i, receipt_id='', invoice_number
                                                                            discount_percent=discount_percent,
                                                                            round_off_total=round_off_total,
                                                                            cess_tax=cess_tax,
-                                                                           overall_discount=overall_discount)
+                                                                           overall_discount=overall_discount,
+                                                                           remarks = remarks)
         seller_received_list.append(
             {'seller_id': '', 'sku_id': data.open_po.sku_id, 'quantity': value,
-             'id': seller_po_summary.id})
+             'id': seller_po_summary.id, 'remarks': remarks})
     else:
         for sell_po in seller_pos:
             if not value:
@@ -2529,10 +2554,11 @@ def update_seller_po(data, value, user, myDict, i, receipt_id='', invoice_number
                                                                                invoice_date=invoice_date,
                                                                                round_off_total=round_off_total,
                                                                                cess_tax=cess_tax,
-                                                                               overall_discount=overall_discount)
+                                                                               overall_discount=overall_discount,
+                                                                               remarks = remarks)
             seller_received_list.append(
                 {'seller_id': sell_po.seller_id, 'sku_id': data.open_po.sku_id, 'quantity': value,
-                 'id': seller_po_summary.id})
+                 'id': seller_po_summary.id, 'remarks': remarks})
     return seller_received_list
 
 
@@ -2557,7 +2583,7 @@ def create_file_po_mapping(request, user, receipt_no, myDict):
                                       master_type='PO_TEMP')
             if exist_master_docs:
                 for exist_master_doc in exist_master_docs:
-                    if os.path.exists(exist_master_doc.uploaded_file):
+                    if exist_master_doc.uploaded_file and os.path.exists(exist_master_doc.uploaded_file.path):
                         os.remove(exist_master_doc.uploaded_file.path)
                     exist_master_doc.delete()
     except Exception as e:
@@ -2565,6 +2591,36 @@ def create_file_po_mapping(request, user, receipt_no, myDict):
         log.debug(traceback.format_exc())
         log.info("Create GRN File Mapping failed for user " + str(user.username) + \
                  " and error statement is " + str(e))
+
+
+def update_remarks_put_zone(remarks, user, put_zone, seller_summary_id=''):
+    if remarks == 'offer_applied':
+        offer_zone_check = ZoneMaster.objects.filter(zone='Offer Change', user=user.id)
+        if offer_zone_check.exists():
+            put_zone = offer_zone_check[0].zone
+    elif remarks == 'mrp_change':
+        mrp_change_check = ZoneMaster.objects.filter(zone='MRP Change', user=user.id)
+        if mrp_change_check.exists():
+            put_zone = mrp_change_check[0].zone
+    elif remarks == 'mrp_change,offer_applied':
+        offer_zone_check = ZoneMaster.objects.filter(zone='Offer Change', user=user.id)
+        if offer_zone_check.exists():
+            put_zone = offer_zone_check[0].zone
+    elif seller_summary_id:
+        seller_po_summary = SellerPOSummary.objects.filter(id=seller_summary_id)
+        if seller_po_summary:
+            seller_po_summary = seller_po_summary[0]
+            if seller_po_summary.batch_detail and seller_po_summary.purchase_order.open_po:
+                other_mrp_stock = StockDetail.objects.filter(sku__user=user.id, quantity__gt=0,
+                                                             sku_id=seller_po_summary.purchase_order.open_po.sku_id,
+                                           sellerstock__seller_id=seller_po_summary.seller_po.seller_id).\
+                    exclude(Q(location__zone__zone__in=get_exclude_zones(user)) |
+                              Q(batch_detail__mrp=seller_po_summary.batch_detail.mrp))
+                if other_mrp_stock.exists():
+                    mrp_change_check = ZoneMaster.objects.filter(zone='MRP Change', user=user.id)
+                    if mrp_change_check.exists():
+                        put_zone = mrp_change_check[0].zone
+    return put_zone
 
 
 def generate_grn(myDict, request, user, is_confirm_receive=False):
@@ -2732,7 +2788,10 @@ def generate_grn(myDict, request, user, is_confirm_receive=False):
                 put_zone = put_zone[0]
 
             put_zone = put_zone.zone
-
+        if seller_received_list:
+            remarks = seller_received_list[0].get('remarks', '')
+            put_zone = update_remarks_put_zone(remarks, user, put_zone,
+                                               seller_summary_id=seller_received_list[0].get('id', ''))
         temp_dict = {'received_quantity': float(value), 'user': user.id, 'data': data, 'pallet_number': pallet_number,
                      'pallet_data': pallet_data}
 
@@ -4099,14 +4158,17 @@ def check_wms_qc(request, user=''):
                                     'options': REJECT_REASONS, 'use_imei': use_imei}))
 
 
-def get_seller_received_list(data, user):
+def get_seller_received_list(data, user, sps_obj=None):
     seller_po_summary = SellerPOSummary.objects.filter(Q(location_id__isnull=True) | Q(location__zone__zone='QC_ZONE'),
                                                        seller_po__seller__user=user.id, putaway_quantity__gt=0,
                                                        seller_po__open_po_id=data.open_po_id)
+    if sps_obj:
+        seller_po_summary = seller_po_summary.filter(id=sps_obj.id)
     seller_received_list = []
     for summary in seller_po_summary:
         seller_received_list.append({'seller_id': summary.seller_po.seller_id, 'sku': summary.seller_po.open_po.sku,
-                                     'quantity': summary.putaway_quantity, 'id': summary.id})
+                                     'quantity': summary.putaway_quantity, 'id': summary.id,
+                                     'remarks': summary.remarks })
     return seller_received_list
 
 
@@ -6428,10 +6490,12 @@ def get_segregation_pos(start_index, stop_index, temp_data, search_term, order_t
 def get_po_segregation_data(request, user=''):
     segregations = get_primary_suggestions_data(request, user)
     order_id = request.GET['order_id']
+
     segregations = segregations.select_related('purchase_order', 'batch_detail').\
                                 filter(purchase_order__order_id=order_id)
     if not segregations:
         return HttpResponse("No Data found")
+
     po_reference = get_po_reference(segregations[0].purchase_order)
     orders = []
     order_data = {}
@@ -6462,6 +6526,16 @@ def get_po_segregation_data(request, user=''):
                 po_tax = open_po.cgst_tax + open_po.sgst_tax + open_po.igst_tax + open_po.cess_tax + open_po.utgst_tax
                 if seg_tax != po_tax:
                     deviation_remarks.append('Tax Rate Deviation')
+            if segregation_obj.seller_po_summary:
+                seller_po_summary_obj = segregation_obj.seller_po_summary
+                remarks = seller_po_summary_obj.remarks
+                if str(remarks).find('offer_applied') != -1 :
+                    offer_check = True
+                else:
+                    offer_check = False
+            else:
+                offer_check = False
+
 
         order = segregation_obj.purchase_order
         order_data = get_purchase_order_data(order)
@@ -6478,6 +6552,7 @@ def get_po_segregation_data(request, user=''):
             data_dict = {'segregation_id': segregation_obj.id,'order_id': order.id, 'wms_code': order_data['wms_code'],
                             'sku_desc': order_data['sku_desc'],
                             'quantity': quantity, 'sellable': quantity,
+                            'offer_check' :offer_check,
                             'non_sellable': 0,
                             'name': str(order.order_id) + '-' + str(
                                 re.sub(r'[^\x00-\x7F]+', '', order_data['wms_code'])),
@@ -6539,9 +6614,29 @@ def confirm_primary_segregation(request, user=''):
             segregation_obj.status = 0
             segregation_obj.save()
             batch_dict = {}
+            if segregation_obj.seller_po_summary:
+                seller_po_summary_obj = segregation_obj.seller_po_summary
+                remarks = seller_po_summary_obj.remarks
+                remarks = str(remarks)
+                offer_applicable = data_dict['offer_applicable'][ind]
+                remark_keys = []
+                if remarks:
+                    remark_keys = remarks.split(',')
+                if offer_applicable == 'true':
+                    if 'offer_applied' not in remark_keys:
+                        remark_keys.append('offer_applied')
+                elif offer_applicable == 'false':
+                    if 'offer_applied' in remark_keys:
+                        remark_keys.remove('offer_applied')
+                remarks = ','.join(remark_keys)
+                seller_po_summary_obj.remarks = remarks
+                seller_po_summary_obj.save()
+
             if segregation_obj.batch_detail:
                 batch_detail = segregation_obj.batch_detail
                 manufactured_date = ''
+
+
                 if batch_detail.manufactured_date:
                     manufactured_date = batch_detail.manufactured_date.strftime('%m/%d/%Y')
                 expiry_date = ''
@@ -6554,7 +6649,8 @@ def confirm_primary_segregation(request, user=''):
                               'mrp': batch_detail.mrp, 'buy_price': batch_detail.buy_price,
                               'weight': batch_detail.weight, 'ean_number': batch_detail.ean_number}
             purchase_data = get_purchase_order_data(segregation_obj.purchase_order)
-            seller_received_dict = get_seller_received_list(segregation_obj.purchase_order, user)
+            seller_received_dict = get_seller_received_list(segregation_obj.purchase_order, user,
+                                                            sps_obj=segregation_obj.seller_po_summary)
             if sellable:
                 put_zone = purchase_data['zone']
                 if put_zone:
@@ -6571,6 +6667,10 @@ def confirm_primary_segregation(request, user=''):
                              'pallet_number': '', 'pallet_data': {}}
                 seller_received_dict, seller_summary_dict = get_quality_check_seller(seller_received_dict, temp_dict,
                                                                                      purchase_data)
+                for seller_index, seller_received in enumerate(seller_received_dict):
+                    put_zone = update_remarks_put_zone(seller_received.get('remarks', ''), user, put_zone,
+                                                       seller_summary_id=seller_received.get('id', ''))
+                    seller_received_dict[seller_index]['put_zone'] = put_zone
                 save_po_location(put_zone, temp_dict, seller_received_list=seller_summary_dict, run_segregation=False,
                                  batch_dict=batch_dict)
             sellable_qty = get_decimal_limit(user.id, (float(segregation_obj.sellable) + sellable))
@@ -6587,6 +6687,10 @@ def confirm_primary_segregation(request, user=''):
                              'pallet_number': '', 'pallet_data': {}}
                 seller_received_dict, seller_summary_dict = get_quality_check_seller(seller_received_dict, temp_dict,
                                                                                      purchase_data)
+                for seller_index, seller_received in enumerate(seller_received_dict):
+                    put_zone = update_remarks_put_zone(seller_received.get('remarks', ''), user, put_zone,
+                                                       seller_summary_id=seller_received.get('id', ''))
+                    seller_received_dict[seller_index]['put_zone'] = put_zone
                 save_po_location(put_zone, temp_dict, seller_received_list=seller_summary_dict, run_segregation=False,
                                  batch_dict=batch_dict)
             non_sellable_qty = get_decimal_limit(user.id, (float(segregation_obj.non_sellable) + non_sellable))
